Reduce parsing runtime by around 27%
Our current scoring model is a probabilistic model that adjusts the
probability of a solution by multiplying it by other probabilities.
Since a probability is at most 1, a partial score in a probabilistic
model can never increase; in practice, it will only decrease as more and
more terms are added.

In other words, a solution's score can only decrease as we add more
words to it.

If we see a complete solution in our priority queue with score C, we
thus know that all solutions following it both (a) have a score equal or
lower to C due to the nature of priority queues and (b) cannot possibly
produce a result better than C per our result above.

So once we find a finished solution in our priority queue, we can
suspend execution.

Before:

    18.91 s  100.0% 927.00 ms   vidyut::segmenting::segment::h6a19c09ae9c14b0f

After:

    13.79 s  100.0% 657.00 ms   vidyut::segmenting::segment::h6a19c09ae9c14b0f