[lipi] Add a stronger implementation
This commit adds support for a variety of schemes as defined by the
`indic_transliteration` project. It also adds a minimal test suite as a
stronger guarantee on program correctness.

`vidyut-lipi` is still immature compared to transliterators like
Aksharamukha or `indic_transliteration`. However, it is on a good
trajectory, and I think it will become a compelling transliteratior
backend over time.

Some notes on design:

- This commit does not use any of the code from @skmnktl's in-progress
  transliterator, but it does borrow the idea of incorporating
  `indic_transliteration`'s TOML maps directly into the program source
  code.

- I liked @skmnktl's idea of using a `Token` enum as an intermediate
  representation between the input scheme and the output scheme, but
  pursing that approach felt cumbersome when mapping between *sequences*
  of characters (e.g. when working with ITRANS), so I've stayed with the
  approach used by `indic_transliteration`, i.e. using Devanagari as the
  intermediate representation.