[prakriya] Use the full Ganapatha with metadata
This resolves GitHub issues #160, #161, and #162.

CONTEXT

Various rules in the Ashtadhyayi refer to one or more *ganas*,
collections of stems and other linguistic data that enumerate the
items that can trigger the rule.

Our previous approach used a small number of manually curated ganas
that we then used in the appropriate rules. But this gana data was
ultimately just copied from the data on ashtadhyayi.com, perhaps with
a few tweaks along the way.

Given this partial treatment, I've felt that it would be worthwhile
to properly and maturely model all of these ganas.

TECHNICAL DECISIONS

Given the number and size of these ganas, it seems simplest to couple
tightly with ashtadhyayi.com's data as a source of truth then derive the
API we want programmatically. If there are errors in the upstream,
fixing them will fix our data as well and improve the ecosystem overall.

The main technical decision here was this: do we store the data in the
code itself, or do we rely instead on an external data file?

Both positions have merits. Using an external data file means that users
can specify precisely which gana they want to use when passing some
linguistic item into the system. Doing so also follows the pattern we
use for the Dhatupatha, which is likewise stored as an external file.

Despite these merits, I opted to store the Ganapatha data directly in
the code itself. I am not convinced that this approach is right, but I
chose it for the following reasons:

- The Ganapatha and the Dhatupatha follow different usage patterns.
  Dhatus are almost always part of exactly one gana, whereas items in
  the Ganapatha are often in multiple ganas.

- Dhatu ganas are common knowledge and familiar, and almost anyone who
  knows some dhatu will also know its gana. Therefore, there is less
  friction in defining the gana explicitly (e.g. through a data file)
  then passing it in. This is not the case for items in the Ganapatha.

- My hunch is that there is generally more variance in Dhatupathas,
  which motivates giving the user extra control by modeling it as an
  external file. I don't think this is the case for the Ganapatha.