[kosha] Support tinanta prefixes
This is a space optimization that reduces the total size of the kosha.
I haven't benchmarked lookup time, but given that our main use case now
seems to be interactive lookup, times below 1ms are good enough for the
time being.

Before:       65M FST,  27M registry
After:        44M FST,  29M registry
Difference:  -21M FST,  +2M registry

APPROACH

Our underlying FST excels when keys share common prefixes and suffixes
and when their values can accumulate from character to character. It
tends to bloat when very different kinds of linguistic data share the
same key prefix. When this happens, we lose the ability to encode
information at the base and must shove it in near the suffix.

At least, this is my basic model based on a few experiments building
the kosha with different kinds of data. With this in mind, I tried to
avoid collisions and replication as much as I can and saw an early
win by storing only a krdanta prefix rather than an entire krdanta
itself. So I decided to try this same approach for tinantas, since
early experiments indicated that this could lead to substantial space
savings.

Supporting tinanta prefixes meant creating a new `PartOfSpeech` type,
but since we already had 4 types in use, I could not do so without
assigning an extra bit to track the part of speech. I didn't want to
do that, so I took this opportunity to also make two deprecations:

- `Avyaya` data is now stored on the pratipadika instead.
- `Unknown` is removed, since its main use case is as a convenience
  value for `cheda`.

This setup means we now have four entry types: Subanta, SubantaPrefix,
Tinanta, TinantaPrefix. I think the *Prefix types can be further
combined, but that's out of scope here.

Much of the time sink when creating this commit was in managing the
much more complicated procedure for key lookup and ensuring that we
were calculating node values correctly when traversing the FST. I
haven't benchmarked this approach and don't know how slow it is, but
that's fine for now: the most important thing is to include correct
data.

As part of this commit, I also expanded `debug_kosha` and used it to
find and remove certain patterns in the data that I thought were
buggy or unpleasant. This includes a lot of redundant data that, when
removed, leads to a modest ~2MB reduction in size, in addition to
making the kosha data much more trustworthy.

USE CASES

The zipped dictionary data on disk is 29MB. Based on the size of the
data files fetch by ashtadhyayi.com, I think this is already with range
for client-side usage. But, I think 10MB is a better threshold to aim
for, and I think we can get close with more attention.

The trick is to move as much metadata to the registry as possible,
since it compresses well. Sample `zip` results:

    adding: kosha/registry.msgpack (deflated 82%)
    adding: kosha/padas.fst (deflated 46%)

I'm inclined toward combining prefixes for different kinds of
linguistic data, e.g. mapping `buBUz` to all of its subantas and
tinantas. But more experimentation will surface better ways of
going about this.